Ajout de la bomb
Bombe qui peut être planté + tracker + explosion

+++ src/fr/karamouche/plantthebomb/EventListener.java
@@ -2,9 +2,11 @@ package fr.karamouche.plantthebomb;
 
 import fr.karamouche.plantthebomb.enums.PTBteam;
 import fr.karamouche.plantthebomb.enums.Tools;
+import fr.karamouche.plantthebomb.objects.Bomb;
 import fr.karamouche.plantthebomb.objects.Game;
 import fr.karamouche.plantthebomb.objects.PTBer;
 import org.bukkit.*;
+import org.bukkit.block.Block;
 import org.bukkit.entity.Arrow;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.EntityType;
@@ -22,6 +24,7 @@ import fr.karamouche.plantthebomb.enums.Spawns;
 import fr.karamouche.plantthebomb.enums.Statut;
 import org.bukkit.event.weather.WeatherChangeEvent;
 import org.bukkit.inventory.ItemStack;
+import org.bukkit.scheduler.BukkitRunnable;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -38,16 +41,35 @@ public class EventListener implements Listener {
 	public void onJoin(PlayerJoinEvent event) {
 		myPlugin.getScoreboardManager().onLogin(event.getPlayer());
 		Player player = event.getPlayer();
-		if(myPlugin.getCurrentGame().getStatut().equals(Statut.LOBBY) || myPlugin.getCurrentGame().getStatut().equals(Statut.STARTING)) {
+		Game game = myPlugin.getCurrentGame();
+		if(game.getStatut().equals(Statut.LOBBY)) {
 			player.teleport(Spawns.LOBBY.toLocation());
-			myPlugin.getCurrentGame().addPlayer();
-			event.setJoinMessage(myPlugin.getCurrentGame().getTag()+ ChatColor.YELLOW+event.getPlayer().getName()+ChatColor.AQUA +" a rejoint la partie "+"§8(§a"+myPlugin.getCurrentGame().getNbPlayers()+"§8/§a"+myPlugin.getCurrentGame().getMaxPlayer()+"§8)");
+			game.addPlayer();
+			event.setJoinMessage(game.getTag()+ ChatColor.YELLOW+event.getPlayer().getName()+ChatColor.AQUA +" a rejoint la partie "+"§8(§a"+myPlugin.getCurrentGame().getNbPlayers()+"§8/§a"+myPlugin.getCurrentGame().getMaxPlayer()+"§8)");
 			player.setGameMode(GameMode.SURVIVAL);
 			player.setFoodLevel(20);
 			player.setHealth(20);
 			player.setLevel(0);
 			player.setExp(0);
 			myPlugin.getCurrentGame().giveLobbyItems(player);
+		}else if(game.getStatut().equals(Statut.INGAME)){
+			if(game.getOfflineptbers().containsKey(player.getUniqueId())){
+				PTBer ptber = game.getOfflineptbers().remove(player.getUniqueId());
+				game.getPtbers().put(player.getUniqueId(), ptber);
+				myPlugin.getCurrentGame().getTeam(ptber.getTeam()).addEntry(player.getName());
+				event.setJoinMessage(game.getTag()+ptber.getTeam().getTag()+player.getName()+ChatColor.WHITE+" s'est reconnecté !");
+			}else{
+				player.getInventory().clear();
+				player.setGameMode(GameMode.SPECTATOR);
+				event.setJoinMessage("");
+			}
+		}
+		else{
+			event.setJoinMessage("");
+			if(game.getStatut().equals(Statut.STARTING))
+				player.kickPlayer(game.getTag()+"La partie est complète");
+			else
+				player.kickPlayer(game.getTag()+"La partie est finie");
 		}
 	}
 	
@@ -58,11 +80,33 @@ public class EventListener implements Listener {
 		Player player = event.getPlayer();
 		if(myPlugin.getCurrentGame().getStatut().equals(Statut.LOBBY) || myPlugin.getCurrentGame().getStatut().equals(Statut.STARTING)) {
 			myPlugin.getCurrentGame().removePlayer();
-			event.setQuitMessage(myPlugin.getCurrentGame().getTag()+ChatColor.YELLOW+event.getPlayer().getName()+ChatColor.GRAY+" a quitté la partie "+ChatColor.YELLOW+"["+(myPlugin.getCurrentGame().getNbPlayers())+"/"+myPlugin.getCurrentGame().getMaxPlayer()+"]");
+			event.setQuitMessage(game.getTag()+ChatColor.YELLOW+event.getPlayer().getName()+ChatColor.GRAY+" a quitté la partie "+"§8(§a"+(game.getNbPlayers())+"/§8/§a"+myPlugin.getCurrentGame().getMaxPlayer()+"§8)");
+
 			if(game.getPtbers().containsKey(player.getUniqueId())){
-				game.getPtbers().get(player.getUniqueId()).destroy();
+				game.getPtbers().remove(player.getUniqueId()).destroy();
 			}
-		}
+		}else if(game.getStatut().equals(Statut.INGAME)){
+			if(game.getPtbers().containsKey(player.getUniqueId())){
+				PTBer ptber = game.getPtbers().remove(player.getUniqueId());
+				ptber.kill(null);
+				game.getOfflineptbers().put(player.getUniqueId(), ptber);
+				game.getTeam(ptber.getTeam()).removeEntry(player.getName());
+				if(game.getTeam(ptber.getTeam()).getEntries().size() == 0){
+					if(ptber.getTeam().equals(PTBteam.ANTITERRORISTE)) {
+						game.setScoreA(9);
+						game.getActualRound().winner(PTBteam.ANTITERRORISTE);
+					}
+					else{
+						game.setScoreT(9);
+						game.getActualRound().winner(PTBteam.TERRORISTE);
+					}
+				}
+				event.setQuitMessage(game.getTag()+ptber.getTeam().getTag()+player.getName()+ChatColor.WHITE+" s'est déconnecté !");
+			}else{
+				event.setQuitMessage("");
+			}
+		}else
+			event.setQuitMessage("");
 	}
 
 	@EventHandler
@@ -86,7 +130,7 @@ public class EventListener implements Listener {
 				Player victim = (Player) event.getEntity();
 				double damage = event.getFinalDamage();
 				EntityDamageEvent.DamageCause cause = event.getCause();
-				if(victim.getHealth() - damage <= 0 && !cause.equals(EntityDamageEvent.DamageCause.ENTITY_ATTACK)) {
+				if(victim.getHealth() - damage <= 0 && !cause.equals(EntityDamageEvent.DamageCause.ENTITY_ATTACK) && !cause.equals(EntityDamageEvent.DamageCause.PROJECTILE)) {
 					event.setCancelled(true);
 					PTBer ptber = game.getPtbers().get(victim.getUniqueId());
 					ptber.kill(null);
@@ -144,14 +188,69 @@ public class EventListener implements Listener {
 			event.setCancelled(true);
 		}else if(event.getItemDrop().getItemStack().isSimilar(Tools.SHOP.toItem()))
 			event.setCancelled(true);
+		else if (event.getItemDrop().getItemStack().isSimilar(Tools.BOMB.toItem())){
+			BukkitRunnable run = new BukkitRunnable() {
+				@Override
+				public void run() {
+					game.getActualRound().getBomb().setLoc(event.getItemDrop().getLocation());
+					for(PTBer ptber : game.getPtbers().values()){
+						if(ptber.getTeam().equals(PTBteam.TERRORISTE))
+							game.getActualRound().getBomb().sendDirection(Bukkit.getPlayer(ptber.getPlayerID()));
+					}
+				}
+			};
+			run.runTaskLaterAsynchronously(myPlugin, 20);
+		}
 	}
+
+	@EventHandler
+	public void onPickupItem(PlayerPickupItemEvent event){
+		Player player = event.getPlayer();
+		Game game = myPlugin.getCurrentGame();
+		ItemStack item = event.getItem().getItemStack();
+		if(game.getPtbers().containsKey(player.getUniqueId())){
+			PTBer ptber = game.getPtbers().get(player.getUniqueId());
+			if(item.isSimilar(Tools.BOMB.toItem())){
+				if(ptber.getTeam().equals(PTBteam.TERRORISTE)){
+					Bomb bomb = game.getActualRound().getBomb();
+					bomb.setOwner(ptber);
+					bomb.setLoc(null);
+					event.setCancelled(true);
+					player.getInventory().setItem(8, Tools.BOMB.toItem());
+					player.playSound(player.getLocation(), Sound.ITEM_PICKUP, 1, 1);
+					event.getItem().remove();
+				}else
+					event.setCancelled(true);
+			}
+		}
+	}
+
 	@EventHandler
 	public void onPlaceBlock(BlockPlaceEvent event){
 		Game game = myPlugin.getCurrentGame();
 		if(game.getStatut().equals(Statut.LOBBY) || game.getStatut().equals(Statut.STARTING)){
 			if(!event.getPlayer().isOp() || !event.getPlayer().getGameMode().equals(GameMode.CREATIVE))
 				event.setCancelled(true);
-		}
+		}else if(game.getStatut().equals(Statut.INGAME)){
+			Material blockType = event.getBlock().getType();
+			if(blockType.equals(Material.REDSTONE_TORCH_ON)) {
+				if(!game.getActualRound().isFinish()) {
+					Block downBlock = event.getBlockAgainst();
+					Bomb bomb = game.getActualRound().getBomb();
+					if(downBlock.getType().equals(Material.PRISMARINE)) {
+						bomb.tryplant(event.getPlayer(), event.getBlockPlaced());
+					}
+					else {
+						event.setCancelled(true);
+						event.getPlayer().sendMessage(game.getTag()+ChatColor.RED+"La bombe ne peut être placée que sur un site !");
+						event.getPlayer().playSound(event.getPlayer().getLocation(), Sound.VILLAGER_NO, 1, 1);
+					}
+				}
+				else
+					event.setCancelled(true);
+			}
+		}else
+			event.setCancelled(true);
 	}
 
 	@EventHandler
@@ -239,6 +338,11 @@ public class EventListener implements Listener {
 				player.sendMessage(game.getTag() + ChatColor.RED + "La partie n'a pas commencée");
 				player.playSound(event.getPlayer().getLocation(), Sound.VILLAGER_NO, 1, 1);
 			}
+		}else if(game.getStatut().equals(Statut.INGAME) && game.getPtbers().containsKey(player.getUniqueId())){
+			PTBer ptber = myPlugin.getCurrentGame().getPtbers().get(player.getUniqueId());
+			if(ptber.getTeam().equals(PTBteam.TERRORISTE) && game.getActualRound().getBomb().isDrop()){
+				game.getActualRound().getBomb().sendDirection(player);
+			}
 		}
 	}
 
@@ -254,4 +358,52 @@ public class EventListener implements Listener {
 	public void onWeather(WeatherChangeEvent event){
 		event.setCancelled(true);
 	}
+
+	@EventHandler
+	public void onChat(AsyncPlayerChatEvent event){
+		Player player = event.getPlayer();
+		String message = event.getMessage();
+		Game game = myPlugin.getCurrentGame();
+		if(game.getPtbers().containsKey(player.getUniqueId())) {
+			PTBer ptber = game.getPtbers().get(player.getUniqueId());
+			if(message.startsWith("!")){
+				message = message.substring(1);
+				if(ptber.getTeam().equals(PTBteam.ANTITERRORISTE))
+					event.setFormat("§b[Antiterroriste] "+player.getName()+"§r: "+message);
+				else
+					event.setFormat("§c[Terroriste] "+player.getName()+"§r: "+message);
+			}else{
+				event.setCancelled(true);
+				for(PTBer ptbers : game.getPtbers().values()){
+					if(ptbers.getTeam().equals(ptber.getTeam())){
+						Player players = Bukkit.getPlayer(ptbers.getPlayerID());
+						players.sendMessage(player.getName()+" -> "+ptber.getTeam().toString()+ ChatColor.WHITE+": "+message);
+					}
+				}
+			}
+		}else{
+			event.setCancelled(true);
+			for(Player players : Bukkit.getOnlinePlayers()){
+				if(!game.getPtbers().containsKey(players.getUniqueId()))
+					players.sendMessage(ChatColor.GRAY+"[Spectateur] "+player.getName()+": "+message);
+			}
+		}
+	}
+
+	/*public void onChat(AsyncPlayerChatEvent event) {
+		Player player = event.getPlayer();
+		String message = event.getMessage();
+		Team playerTeam = FPTB.board.getPlayerTeam(player);
+		if(playerTeam != null) {
+			if(playerTeam.equals(FPTB.terroriste)) {
+				event.setFormat("§c[Terroriste] "+player.getName()+"§r: "+message);
+			}
+			else if(playerTeam.equals(FPTB.antiterroriste)) {
+				event.setFormat("§b[Antiterroriste] "+player.getName()+"§r: "+message);
+			}
+			else if(playerTeam.equals(FPTB.spectator)) {
+				event.setFormat("§8[Spectateur] "+player.getName()+"§r: "+message);
+			}
+		}
+	}*/
 }
